requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Data Science
numpy>=1.24.0
matplotlib>=3.7.0

# Utilities
tqdm>=4.65.0

# Jupyter
jupyter>=1.0.0
